feat: use assets
