+++ src/hooks/index.ts
@@ -7,3 +7,4 @@ export * from './fuzzyMatch';
 export * from './queue';
 export * from './networkColor';
 export * from './records';
+export * from './assets';
